+++ IntegrationBus/source/extensions/Tests/IbExtensionsTest.cpp
@@ -2,23 +2,61 @@
 
 #include "IbExtensions.hpp"
 #include "ib/version.hpp"
+#include "ib/cfg/ExtensionConfigBuilder.hpp"
 #include <tuple>
 
 #include "DummyExtension.hpp"
 
 #include "gtest/gtest.h"
 #include "gmock/gmock.h"
-#include <cstdlib>
+
+#if defined(WIN32)
+// without underscore is deprecated on windows
+#define getcwd _getcwd
+#define chdir _chdir
+#endif
 
 using namespace testing;
+
+namespace
+{
+    std::string GetCurrentWorkingDir()
+    {
+        constexpr size_t maxPath = 4096;
+        char buffer[maxPath];
+        if (getcwd(buffer, maxPath) == nullptr)
+        {
+            throw std::runtime_error("Couldn't get current working directory.");
+        }
+        return std::string(buffer);
+    }
+    void SetCurrentWorkingDir(const std::string& cwd)
+    {
+        if (chdir(cwd.c_str()) != 0)
+        {
+            throw std::runtime_error("Couldn't set the current working directory.");
+        }
+    }
+}
 
 class IbExtensionsTest : public Test
 {
 protected:
-    IbExtensionsTest()
+    void TearDown() override
     {
+        // Switch to original directory if the previous test fails
+        SetCurrentWorkingDir(currentWorkingDir);
     }
+
+    static void SetUpTestCase()
+    {
+        currentWorkingDir = GetCurrentWorkingDir();
+    }
+
+    static std::string currentWorkingDir;
 };
+
+std::string IbExtensionsTest::currentWorkingDir;
 
 using triple = std::tuple<uint32_t, uint32_t, uint32_t>;
 
@@ -49,6 +87,26 @@ TEST_F(IbExtensionsTest, load_dummy_lib)
 
     }
     //must not crash when going out of scope
+}
+
+// Use the search path hints specified in the extension configuration
+TEST_F(IbExtensionsTest, load_dummy_lib_from_custom_search_path)
+{
+    try
+    {
+        SetCurrentWorkingDir("..");
+
+        ib::cfg::ExtensionConfig config;
+        EXPECT_THROW(ib::extensions::LoadExtension("DummyExtension", config), ib::extensions::ExtensionError);
+
+        config.searchPathHints.emplace_back(currentWorkingDir);
+        ASSERT_NE(ib::extensions::LoadExtension("DummyExtension", config), nullptr);
+    }
+    catch (std::exception& e)
+    {
+        std::cout << e.what() << '\n';
+        FAIL();
+    }
 }
 
 TEST_F(IbExtensionsTest, dynamic_cast)
@@ -108,12 +166,15 @@ TEST_F(IbExtensionsTest, multiple_extensions_loaded)
     mod2->SetDummyValue(1337);
     EXPECT_NE(mod1->GetDummyValue(), 1337);
 }
+
 #if !defined(_WIN32)
 TEST_F(IbExtensionsTest, load_from_envvar)
 {
-    setenv("TEST_VAR", "../Tests", 1); //should be invariant
-    std::vector<std::string> hints={"ENV:TEST_VAR"};
-    auto base1 = ib::extensions::LoadExtension("DummyExtension", hints);
+    SetCurrentWorkingDir("..");
+    setenv("TEST_VAR", "Tests", 1); //should be invariant
+    ib::cfg::ExtensionConfig config;
+    config.searchPathHints.emplace_back("ENV:TEST_VAR");
+    auto base1 = ib::extensions::LoadExtension("DummyExtension", config);
     auto* mod1 = dynamic_cast<DummyExtension*>(base1.get());
     mod1->SetDummyValue(1);
     EXPECT_EQ(mod1->GetDummyValue(), 1);